Add files via upload
Made in the COMPETENCE CENTER IN MANUFACTURING (CCM), a laboratory of the AERONAUTICS INSTITUTE OF TECHNOLOGY (ITA). This dataset was produced within a Master's Dissertation work consisting of experiments based on turning an AISI H13 steel with cutting fluid. The experiment with the cutting tool flank wear varying in three levels produced 288 samples for each measured roughness parameter. The experiment used the tool condition as a factor, bringing the necessity of having the tool worn to different levels---new tool (VBB = 0.0 mm), mid-life tool (VBB = 0.1 mm), and end-of-life tool (VBB = 0.3 mm). We will use this experiment to build ANN models to predict the arithmetic mean deviation (Ra)